+++ client/src/main/kotlin/dev/voroscsoki/stratopolis/client/api/HttpAccessor.kt
@@ -1,0 +1,13 @@
+package dev.voroscsoki.stratopolis.client.api
+import java.net.URI
+import java.net.URL
+import java.net.http.HttpClient
+
+class HttpAccessor {
+    companion object {
+
+        fun testRequest() : String {
+            return URI("http://localhost:8085/test").toURL().readText().toString()
+        }
+    }
+}
