+++ server/build.gradle.kts
@@ -2,6 +2,7 @@ val mainName = "dev.voroscsoki.stratopolis.server.Main"
 plugins {
     kotlin("jvm")
     application
+    id("io.ktor.plugin") version "2.3.12"
 }
 
 application {
@@ -16,6 +17,12 @@ repositories {
     }
     maven {
         url = uri("https://mvn.slimjars.com/")
+    }
+    maven {
+        url = uri("https://maven.pkg.jetbrains.space/public/p/ktor/eap")
+    }
+    maven {
+        url = uri("https://maven.pkg.jetbrains.space/kotlin/p/kotlin/dev")
     }
 }
 
@@ -32,6 +39,11 @@ dependencies {
     implementation("de.topobyte:osm4j-incubating:1.3.0")
     implementation("de.topobyte:osm4j-replication:1.3.0")
     implementation("de.topobyte:osm4j-testing:1.3.0")
+
+    implementation("io.ktor:ktor-server-core")
+    implementation("io.ktor:ktor-server-netty")
+    implementation("io.ktor:ktor-server-config-yaml")
+    implementation(project(":common"))
 }
 
 tasks.register("buildBackend", GradleBuild::class) {
@@ -41,6 +53,6 @@ tasks.register("buildBackend", GradleBuild::class) {
 tasks.register("runBackend", JavaExec::class) {
     group = "application"
     mainClass.set(this@Build_gradle.mainName)
-    classpath = sourceSets["main"].runtimeClasspath
+    classpath = sourceSets["main"].compileClasspath
 }
