+++ server/src/main/kotlin/osm/OsmStorage.kt
@@ -2,11 +2,12 @@ package dev.voroscsoki.stratopolis.server.osm
 
 import de.topobyte.osm4j.core.model.iface.*
 import de.topobyte.osm4j.pbf.seq.PbfIterator
-
-typealias CoordPair = Pair<Double, Double>
+import dev.voroscsoki.stratopolis.common.api.Building
 
 class OsmStorage(source: PbfIterator) {
-    val buildings = mutableListOf<Building>()
+    companion object {
+        val buildings = mutableListOf<Building>()
+    }
     val seq = source.iterator().asSequence()
 
     init {
@@ -25,18 +26,4 @@ class OsmStorage(source: PbfIterator) {
 
 fun OsmEntity.isBuilding(): Boolean {
     return this.getTags().any { it.key == "building" }
-}
-
-
-class Building(
-    private val id: Long,
-    val tags: List<OsmTag>,
-    val metadata: OsmMetadata,
-    val type: EntityType,
-    val coords: CoordPair = Pair(0.0, 0.0),
-    val lines : List<OsmWay> = emptyList()
-){
-    override fun toString(): String {
-        return "Building(id=$id, tags=$tags, metadata=$metadata, type=$type, coords=$coords, lines=$lines)"
-    }
 }
