+++ settings.gradle.kts
@@ -2,3 +2,4 @@ rootProject.name = "Stratopolis"
 
 include("server")
 include("client")
+include("common")
